# openapi.yaml
openapi: 3.1.1
info:
  title: Open-Spot API
  description: |
    개인 장소 기록 및 관리 서비스 API

    ## 주요 기능
    - Google OAuth2 인증 (Bearer Token + HttpOnly Cookie)
    - 개인 장소 기록 관리 (평점, 리뷰, 태그)
    - 장소 그룹 관리
    - 푸시 알림 관리

    ## 인증 방식
    - **Access Token**: Authorization 헤더로 전달 (Bearer 방식)
    - **Refresh Token**: HttpOnly Cookie로 전달 (XSS 방어)

    ## MSA 아키텍처
    - **Gateway (8080)**: 모든 API의 진입점, JWT 검증 수행
    - **Auth Service (8081)**: Google OAuth2 로그인, JWT 발급
    - **Location Service (8082)**: 장소 관리
    - **Notification Service (8083)**: 알림 관리

    **중요**: 프론트엔드는 반드시 Gateway(8080)를 통해 API를 호출해야 합니다.
  version: 1.0.0
  contact:
    name: Open-Spot Team
servers:
  - url: http://localhost:8080
    description: Gateway (모든 API 호출은 이 URL 사용)
  - url: http://localhost:8081
    description: Auth Service (직접 호출 금지, Gateway 경유 필수)
  - url: http://localhost:8082
    description: Location Service (직접 호출 금지, Gateway 경유 필수)
  - url: http://localhost:8083
    description: Notification Service (직접 호출 금지, Gateway 경유 필수)
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/auth/google/login:
    get:
      tags:
        - Authentication
      summary: Google OAuth2 로그인 시작
      description: |
        Google OAuth2 로그인 플로우를 시작합니다.
        로그인 성공 시 프론트엔드로 리다이렉트하며, Access Token은 URL 쿼리 파라미터로, Refresh Token은 HttpOnly 쿠키로 전달됩니다.

        **리다이렉트 예시**:
        `http://localhost:3000/auth/login/success?access_token=eyJhbGc...&new_user=false`

        **프론트엔드 처리**:
        1. URL 쿼리에서 `access_token` 추출
        2. 메모리 또는 localStorage에 저장
        3. 이후 API 요청 시 `Authorization: Bearer <access_token>` 헤더 사용
      operationId: googleLogin
      parameters:
        - name: redirect_uri
          in: query
          required: false
          schema:
            type: string
            format: uri
          description: 로그인 성공 후 리다이렉트할 URL (화이트리스트에 등록된 URL만 허용)
      security: []
      responses:
        '302':
          description: |
            Google 로그인 페이지로 리다이렉트 → 로그인 성공 시 프론트엔드로 리다이렉트
            - Access Token: URL 쿼리 파라미터 (`?access_token=xxx&new_user=false`)
            - Refresh Token: HttpOnly Cookie
          headers:
            Set-Cookie:
              description: |
                Refresh Token만 HttpOnly 쿠키로 설정됩니다:
                - refresh_token: JWT Refresh Token (7일, HttpOnly, Secure, SameSite=Strict)
              schema:
                type: string
                example: refresh_token=eyJhbGc...; HttpOnly; Secure; Path=/; SameSite=Strict; Max-Age=604800
            Location:
              description: 프론트엔드 리다이렉트 URL (Access Token 포함)
              schema:
                type: string
                example: http://localhost:3000/auth/login/success?access_token=eyJhbGc...&new_user=false
  /api/v1/auth/token/refresh:
    post:
      tags:
        - Authentication
      summary: JWT 토큰 갱신
      description: |
        HttpOnly 쿠키의 Refresh Token을 사용하여 새로운 Access Token을 발급받습니다.

        **요청**: Refresh Token은 자동으로 쿠키에서 읽힘
        **응답**:
        - Access Token: Response Body (프론트엔드에서 메모리 저장)
        - Refresh Token: HttpOnly Cookie (자동 갱신)

        **프론트엔드 처리**:
        1. Response Body에서 `accessToken` 추출
        2. 메모리 또는 localStorage에 저장
        3. 이후 API 요청 시 `Authorization: Bearer <accessToken>` 헤더 사용
      operationId: refreshToken
      security: []
      responses:
        '200':
          description: 토큰 갱신 성공 (Access Token은 Response Body, Refresh Token은 HttpOnly 쿠키)
          headers:
            Set-Cookie:
              description: |
                갱신된 Refresh Token만 HttpOnly 쿠키로 설정됩니다:
                - refresh_token: 새로운 Refresh Token (7일, HttpOnly, Secure, SameSite=Strict)
              schema:
                type: string
                example: refresh_token=eyJhbGc...; HttpOnly; Secure; Path=/; SameSite=Strict; Max-Age=604800
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TokenRefreshResponse'
              example:
                success: true
                data:
                  accessToken: eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiI...
                  accessTokenExpiresIn: 3600000
                  message: Token refreshed successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/logout:
    post:
      tags:
        - Authentication
      summary: 로그아웃
      description: |
        Bearer Token을 검증하고 Redis의 Refresh Token을 무효화합니다.
        HttpOnly Cookie의 Refresh Token도 삭제됩니다.

        **요청**: `Authorization: Bearer <access-token>` 헤더 필요
        **응답**: Refresh Token 쿠키 삭제
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 로그아웃 성공 (Refresh Token 쿠키가 삭제됨)
          headers:
            Set-Cookie:
              description: |
                Refresh Token 쿠키가 삭제됩니다 (Max-Age=0):
                - refresh_token
              schema:
                type: string
                example: refresh_token=; HttpOnly; Secure; Path=/; SameSite=Strict; Max-Age=0
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LogoutResponse'
              example:
                success: true
                data:
                  message: Successfully logged out
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ========================================
  # Users APIs
  # ========================================
  /api/v1/users/self:
    get:
      tags:
        - Users
      summary: 현재 사용자 프로필 조회
      description: |
        Gateway에서 검증된 JWT의 사용자 ID로 프로필을 조회합니다.

        **인증 플로우**:
        1. 클라이언트 → Gateway: `Authorization: Bearer <token>`
        2. Gateway: JWT 검증 후 X-User-Id 헤더 추가
        3. Auth Service: X-User-Id 헤더로 사용자 식별
      operationId: getCurrentUser
      responses:
        '200':
          description: 사용자 프로필 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/users/{userId}:
    get:
      tags:
        - Users
      summary: 사용자 프로필 조회
      description: 특정 사용자 ID로 프로필을 조회합니다. (관리자 또는 친구 기능용)
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: 사용자 프로필 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ========================================
  # Categories APIs
  # ========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Access Token을 Authorization 헤더로 전달합니다.

        **사용법**: `Authorization: Bearer <access-token>`

        **토큰 획득**:
        1. Google 로그인 성공 시 URL 쿼리 파라미터로 전달
        2. 토큰 갱신 API 응답 본문에서 획득

        **인증 플로우 (MSA 표준)**:
        1. 클라이언트 → Gateway: `Authorization: Bearer <token>` 헤더 포함
        2. Gateway: JWT 검증 (HS256 알고리즘)
        3. Gateway: 검증 성공 시 `X-User-Id` 헤더 추가
        4. 내부 서비스: `X-User-Id` 헤더로 사용자 식별 (JWT 검증 불필요)
  parameters:
    UserIdParam:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        data:
          type: object
          nullable: true
        error:
          type: object
          nullable: true
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
...
    TokenRefreshResponse:
      type: object
      required:
        - accessToken
        - accessTokenExpiresIn
        - message
      properties:
        accessToken:
          type: string
          description: 새로 발급된 JWT Access Token (프론트엔드 메모리에 저장)
          example: eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiI5ZjY3YzJmZi0zYjJhLTQ5NzUtYjEyZi1lNmM4ZDQwYjk3YjEiLCJleHAiOjE3MzA1NzI3MDAsImlhdCI6MTczMDU2OTEwMH0.xxx
        accessTokenExpiresIn:
          type: integer
          format: int64
          description: 액세스 토큰 만료 시간 (밀리초)
          example: 3600000
        message:
          type: string
          description: 토큰 갱신 성공 메시지
          example: Token refreshed successfully
      description: |
        토큰 갱신 응답.
        - Access Token: Response Body (프론트엔드에서 메모리/localStorage 저장)
        - Refresh Token: HttpOnly Cookie (자동 갱신)
    LogoutResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Successfully logged out

    # ========================================
    # User Schemas
    # ========================================
    UserResponse:
      type: object
      required:
        - id
        - socialId
        - email
        - name
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        socialId:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        pictureUrl:
          type: string
          format: uri
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ========================================
    # Category Schemas
    # ========================================
...